優化寫法

+++ src/main/java/com/worldline/interview/Engine.java
@@ -1,12 +1,14 @@
 package com.worldline.interview;
 
+import java.math.BigDecimal;
+
 public interface Engine {
     void start();
     void stop();
     boolean isRunning();
     void fill(FuelType fuelType, int fuelLevel);
     int getBatchSize();
-    double getCostPerBatch();
+    BigDecimal getCostPerBatch();
     FuelType getFuelType();
     EngineType getEngineType();
 }
